Added common functions to make endpoint more clear
