+++ common/cross-chain/src/events.rs
@@ -1,4 +1,4 @@
-use operation::OperationData;
+use operation::{EventPayment, OperationData};
 
 multiversx_sc::imports!();
 multiversx_sc::derive_imports!();
@@ -9,7 +9,7 @@ pub trait EventsModule {
     fn deposit_event(
         &self,
         #[indexed] dest_address: &ManagedAddress,
-        #[indexed] tokens: &MultiValueEncoded<MultiValue3<TokenIdentifier, u64, EsdtTokenData>>,
+        #[indexed] tokens: &MultiValueEncoded<EventPayment<Self::Api>>,
         event_data: OperationData<Self::Api>,
     );
